Move machines to own folder
