+++ src/v2/components/UpdateView.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { StateFrom } from "xstate";
-import type { machine } from "../utils/updateMachine";
+import type machine from "../machines/updateMachine";
 
 interface UpdateViewProps {
   state: StateFrom<typeof machine>;
